🔒 Add Bash Script [encrypt.sh]
This is a 1500+ line bash script with an easy to use interface for encrypting files. Some feature:
* Works on files of any size including huge files (100s of gigs).
* Works on Ubuntu and Mac without requiring anything else to be installed.
* Works with Windows if using Windows Subsystem for Linux
* Works with Password or Key (keys are securely generated).
* Includes extensive unit testing as part of the script
* Helpful and useful examples when running the script from terminal

It's was designed for compatibility with the PHP Framework of the site. I link to the main site because it shows screenshots and provides technical details. This page won't change and links back to GitHub (main repository and bash script itself).