Add Lanterna
Lanterna is a Java library allowing you to write easy semi-graphical
user interfaces in a text-only environment, very similar to the C
library curses but with more functionality.

One of the main benefits of lanterna is that it is not dependent on any
native library but runs 100% in pure Java.